requirements.txt
httpx>=0.25.0
aiodns>=3.1.1
beautifulsoup4>=4.12.0
lxml>=5.0.0
networkx>=3.2
tldextract>=5.1.0
rich>=13.7.0
pyyaml>=6.0.1
playwright>=1.40.0
dnspython>=2.4.2
uvloop>=0.19.0
